producer/requirements.txt
pandas
numpy
requests
scikit-learn
tensorflow
ta
python-binance
